code reorganisation and first progress
